+++ src/test/java/example/lspAndDemand/requirementsChecking/RedRequirement.java
@@ -1,15 +1,15 @@
-package example.lsp.requirementsChecking;
+package example.lspAndDemand.requirementsChecking;
 
 import lsp.functions.LSPInfo;
 import lsp.LogisticsSolution;
 import lsp.shipment.Requirement;
 
-public class BlueRequirement implements Requirement{
+public class RedRequirement implements Requirement{
 
 	@Override
 	public boolean checkRequirement(LogisticsSolution solution) {
 		for(LSPInfo info : solution.getInfos()) {
-			if(info instanceof BlueInfo) {
+			if(info instanceof RedInfo) {
 				return true;
 			}
 		}
