+++ src/test/java/testSuite/FTLabTestSuite.java
@@ -49,8 +49,8 @@ import lspShipmentAssignmentTests.CompleteLSPShipmentAssignerTest;
 import lspShipmentTest.CollectionShipmentBuilderTest;
 import lspShipmentTest.CompleteShipmentBuilderTest;
 import lspShipmentTest.DistributionShipmentBuilderTest;
-import example.lsp.requirementsChecking.AssignerRequirementsTest;
-import example.lsp.requirementsChecking.TransferrerRequirementsTest;
+import example.lspAndDemand.requirementsChecking.AssignerRequirementsTest;
+import example.lspAndDemand.requirementsChecking.TransferrerRequirementsTest;
 import solutionElementTests.CollectionElementTest;
 import solutionElementTests.DistributionElementTest;
 import lsp.usecase.FirstReloadElementTest;
